\name{compress}
\alias{compress}
\title{Dummy network compression routine}
\description{Dummy routine.}
\usage{compress(dname)}
\arguments{
  \item{dname}{name of gRaven domain}
}
\value{1}
\author{Peter J. Green, \email{P.J.Green@bristol.ac.uk}}
\examples{
demo('chest',echo=FALSE)
compile(chest)
compress(chest)
}
