\name{hugin.domain}
\alias{hugin.domain}
\title{Create a gRaven domain Object}
\description{Create a gRaven domain.}
\usage{hugin.domain()}
\details{Emulates function of the same name in the \code{RHugin} package by calls to \code{gRain} functions}
\value{character string naming a gRaven domain}
\author{Peter J. Green, \email{P.J.Green@bristol.ac.uk}}
\examples{
demo(chest,package="gRaven")
chest
}
