\name{print.gRvname}
\alias{print.gRvname}
\title{Print method for a gRaven domain name}
\description{Print method for a gRaven domain name}
\usage{\method{print}{gRvname}(x, ...)}
\arguments{
  \item{x}{name of gRaven domain}
  \item{...}{additional arguments to \code{\link{print}}}
}
\value{a NULL value is invisibly returned.}
\author{Peter J. Green, \email{P.J.Green@bristol.ac.uk}}
\examples{
demo(chest)
print(chest)
}
